Split up components some more
